V1.4.4 (#7)
* Save current block share for scoresheet page
* Repair on more than 1 unconfirmed block sorting
* Add unlock miner to block candidates
* Unlocker Reward will be paid for those who unlocks a block
* Add redis insert on earnings that will be used on statistic page
* Update calculation upto 5 decimal points
*  New payment processing using transfer_split and dynamic transfer using relay transfer to get actual transfer fee
* Add  payment id validation
* Payment processing checks on minimum payment for integrated address or address with payment id
* Api now has user block scoresheet and pool block scoresheet
* Using hincrbyfloat to get decimal points
* Finalizing new transaction using split transfer and dynamic transfer fee
* Enable loadBalancing

+++ website_example/js/common.js
@@ -224,7 +224,7 @@ function getCoinDecimalPlaces() {
 // Get readable coins
 function getReadableCoins(coins, digits, withoutSymbol){
     var coinDecimalPlaces = getCoinDecimalPlaces();
-    var amount = parseFloat((parseInt(coins || 0) / lastStats.config.coinUnits).toFixed(digits || coinDecimalPlaces));
+    var amount = parseFloat((parseFloat(coins || 0) / lastStats.config.coinUnits).toFixed(digits || coinDecimalPlaces));
     return amount.toString() + (withoutSymbol ? '' : (' ' + lastStats.config.symbol));
 }
